The problem is in the inputLineScene function, used a global event handler for the message event without any mechanism to identify the chat from which the message was received.
This causes the handler for the first call to the function to be saved,
and when a call from another chat room occurs,
that handler is also activated
and the data from one chat room gets into the other.

To fix this problem, added a chat identification mechanism
and used it to filter events accordingly so that only messages
from the chat for which the function was called are handled.